Changed `SourceRectangle` to use `i32`

+++ src/context.rs
@@ -237,7 +237,7 @@ impl<'a, 'b, I: Image> AddImage<'a, 'b, ImageContext<'a, 'b, I>, I> for Context<
             view: Borrowed(self.view.get()),
             transform: Borrowed(self.transform.get()),
             image: Value(image),
-            source_rect: Value([0, 0, w, h]),
+            source_rect: Value([0, 0, w as i32, h as i32]),
         }
     }
 }
