+++ src/rectangle_color_context.rs
@@ -200,7 +200,7 @@ impl<'a, 'b, I: Image> AddImage<'a, 'b, ImageRectangleColorContext<'a, 'b, I>, I
             transform: Borrowed(self.transform.get()),
             rect: Borrowed(self.rect.get()),
             image: Value(image),
-            source_rect: Value([0, 0, w, h]),
+            source_rect: Value([0, 0, w as i32, h as i32]),
             color: Borrowed(self.color.get()),
         }
     }
